requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
black==24.8.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
discord.py==2.4.0
exceptiongroup==1.2.2
executing==2.0.1
fonttools==4.53.1
frozendict==2.4.4
frozenlist==1.4.1
html5lib==1.1
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mplfinance==0.12.10b0
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pyzmq==26.2.0
requests==2.32.3
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
yarl==1.9.4
yfinance==0.2.42
